+++ hb-examples/alarm_clock/alarm_clock.ino
@@ -1,5 +1,4 @@
 #include <hexbright.h>
-#include <twi.h>
 #include <Time.h>
 
 hexbright hb;
