+++ hb-examples/numeric_input/numeric_input.ino
@@ -1,5 +1,4 @@
 #include <hexbright.h>
-#include <twi.h>
 
 hexbright hb;
