+++ hb-examples/stroboscope/stroboscope.ino
@@ -3,7 +3,6 @@
 
 #define BUILD_HACK
 #include <hexbright.h>
-#include <twi.h>
 
 hexbright hb;
