+++ programs/action_timer/action_timer.ino
@@ -28,7 +28,6 @@ either expressed or implied, of the FreeBSD Project.
 */
 
 #include <hexbright.h>
-#include <twi.h>
 #include <Time.h>
 
 hexbright hb;
