+++ programs/functional/functional.ino
@@ -29,8 +29,6 @@ either expressed or implied, of the FreeBSD Project.
 
 #include <hexbright.h>
 
-#include <twi.h>
-
 #define OFF_MODE 0
 #define BLINKY_MODE 1
 #define CYCLE_MODE 2
