+++ programs/tactical/tactical.ino
@@ -31,7 +31,6 @@ either expressed or implied, of the FreeBSD Project.
 
 #define BUILD_HACK
 #include <hexbright.h>
-#include <twi.h>
 
 hexbright hb;
