+++ tests/avr_voltage_test/avr_voltage_test.ino
@@ -29,8 +29,6 @@ either expressed or implied, of the FreeBSD Project.
 
 #include <hexbright.h>
 
-#include <twi.h>
-
 // number of milliseconds between updates
 #define OFF_MODE 0
 #define BLINKY_MODE 1
