+++ tests/print_number_test/print_number_test.ino
@@ -28,7 +28,6 @@ either expressed or implied, of the FreeBSD Project.
 */
 
 #include <hexbright.h>
-#include <twi.h>
 
 // Usage notes are in the readme file in this same directory.
 hexbright hb;
